requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arize==7.48.1
arize-phoenix==11.8.0
arize-phoenix-client==1.13.2
arize-phoenix-evals==0.23.0
arize-phoenix-otel==0.12.1
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
deepmerge==2.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
graphql-core==3.2.6
greenlet==3.2.3
grpc-interceptor==0.15.4
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.4.8
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.47.1
networkx==3.5
numpy==2.3.1
ollama==0.5.1
openinference-instrumentation==0.1.35
openinference-semantic-conventions==0.1.21
openllm-client==0.5.7
openllm-core==0.5.7
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pypdf==5.8.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-futures==1.0.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlean.py==3.49.1
starlette==0.47.1
strawberry-graphql==0.270.1
streamlit==1.47.0
streamlit-chat==0.1.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
